fix: 기타 css 수정 및 경로 수정 #27
